Multiple Changes
Game Manager made but not completed

Changed the Spawn function to take in a game object instead of a transform.

Added lines of code to set the golems target variable in spawnpoint.
EnemyBehaviour added a game object variable.

Got rid of the coroutine and made a checklist function to put in update.

SpawnPointManager takes in a Gameobject instead of a Transform.