Game Manger Worked on.
Moved Some of the scripts to designated folder.

Added Rigidbody to the Golem prefab

Game Manger Worked on.
Game Manager
Made Function to Make a Queue to send to the SpawnManager.
Made a Function  to check if there are a max amount of enemies on the screen.

Enemy Behavoir
Added a Damage variable.
Made the Enemy Check the Collision to see if the hit the player and hit them.